added combine tests + added 1 more testcase to flatten + separated tests per tool (flatten/combine) + moved test files into test/ dir
