+++ package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "bin": "./lib/index.js",
   "scripts": {
-    "test": "mocha"
+    "test": "mocha test/*.spec.js"
   },
   "dependencies": {
     "mkdirp": "^0.5.1",
